configurable array base js code gen, default to 1
The change doesn't change current behavior, but if the following code is
added by the user in her project:

     Blockly.JavaScript.ARRAY_BASE = 0;

Then the list blocks will work like javascript (0 based), instead of 1
based as is the current implementation.

The reason for the change is that because of backward compatibility and
other parts of my app that assume 0 based lists I have a current
incompatibility that causes lot of confusion to users, with this change
all my application treats lists as 0 based and avoids confusion and bugs.